backend_fastapi/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
databases[postgresql]==0.7.6
sqlalchemy==2.0.19
asyncpg==0.27.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.3
